+++ packages/image_picker/image_picker_ios/ios/image_picker_ios/Package.swift
@@ -14,11 +14,15 @@ let package = Package(
   products: [
     .library(name: "image-picker-ios", targets: ["image_picker_ios"])
   ],
-  dependencies: [],
+  dependencies: [
+    .package(name: "FlutterFramework", path: "../FlutterFramework")
+  ],
   targets: [
     .target(
       name: "image_picker_ios",
-      dependencies: [],
+      dependencies: [
+        .product(name: "FlutterFramework", package: "FlutterFramework")
+      ],
       exclude: ["include/image_picker_ios-umbrella.h", "include/ImagePickerPlugin.modulemap"],
       resources: [
         .process("Resources")