+++ packages/url_launcher/url_launcher_macos/macos/url_launcher_macos/Package.swift
@@ -15,11 +15,15 @@ let package = Package(
   products: [
     .library(name: "url-launcher-macos", targets: ["url_launcher_macos"])
   ],
-  dependencies: [],
+  dependencies: [
+    .package(name: "FlutterFramework", path: "../FlutterFramework")
+  ],
   targets: [
     .target(
       name: "url_launcher_macos",
-      dependencies: [],
+      dependencies: [
+        .product(name: "FlutterFramework", package: "FlutterFramework")
+      ],
       resources: [
         .process("Resources")
       ]