+++ packages/pigeon/lib/src/objc/objc_generator.dart
@@ -207,6 +207,34 @@ class ObjcHeaderGenerator extends StructuredGenerator<InternalObjcOptions> {
   }
 
   @override
+  void writeConstants(
+    InternalObjcOptions generatorOptions,
+    Root root,
+    Indent indent, {
+    required String dartPackageName,
+  }) {
+    if (root.constants.isEmpty) {
+      return;
+    }
+    indent.newln();
+    for (final Constant constant in root.constants) {
+      addDocumentationComments(indent, constant.documentationComments, _docCommentSpec);
+      final String constantName = _className(generatorOptions.prefix, constant.name);
+      final String type = constant.type.baseName;
+      if (type == 'String') {
+        final String escaped = escapeStringDoubleQuotes(constant.value.toString());
+        indent.writeln('static NSString *const $constantName = @"$escaped";');
+      } else if (type == 'bool') {
+        final boolVal = (constant.value as bool) ? 'YES' : 'NO';
+        indent.writeln('static const BOOL $constantName = $boolVal;');
+      } else {
+        final _ObjcType objcType = _objcTypeForDartType(generatorOptions.prefix, constant.type);
+        indent.writeln('static const $objcType $constantName = ${constant.value};');
+      }
+    }
+  }
+
+  @override
   void writeEnum(
     InternalObjcOptions generatorOptions,
     Root root,
@@ -467,8 +495,6 @@ class ObjcSourceGenerator extends StructuredGenerator<InternalObjcOptions> {
     indent.writeln('@import Flutter;');
     indent.writeln('#endif');
     indent.newln();
-    _writeDeepEquals(indent);
-    _writeDeepHash(indent);
   }
 
   @override
@@ -1005,6 +1031,9 @@ if (self.wrapped == nil) {
     Indent indent, {
     required String dartPackageName,
   }) {
+    _writeDeepEquals(indent);
+    _writeDeepHash(indent);
+    indent.newln();
     if (root.containsHostApi) {
       _writeWrapError(indent);
       indent.newln();