+++ packages/pigeon/pigeons/core_tests.dart
@@ -4,6 +4,11 @@
 
 import 'package:pigeon/pigeon.dart';
 
+const String aStringConstant = 'stringConstantValue';
+const int anIntConstant = 42;
+const double aDoubleConstant = 3.14;
+const bool aBoolConstant = true;
+
 enum AnEnum { one, two, three, fortyTwo, fourHundredTwentyTwo }
 
 // Enums require special logic, having multiple ensures that the logic can be