+++ packages/pigeon/platform_tests/alternate_language_test_plugin/darwin/alternate_language_test_plugin/Sources/alternate_language_test_plugin/include/alternate_language_test_plugin/CoreTests.gen.h
@@ -14,6 +14,11 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+static NSString *const FLTaStringConstant = @"stringConstantValue";
+static const NSInteger FLTanIntConstant = 42;
+static const double FLTaDoubleConstant = 3.14;
+static const BOOL FLTaBoolConstant = YES;
+
 typedef NS_ENUM(NSUInteger, FLTAnEnum) {
   FLTAnEnumOne = 0,
   FLTAnEnumTwo = 1,