+++ packages/pigeon/platform_tests/alternate_language_test_plugin/example/ios/RunnerTests/DataClassMethodsTest.m
@@ -253,4 +253,11 @@
   XCTAssertEqual(a.hash, b.hash);
 }
 
+- (void)testConstants {
+  XCTAssertEqualObjects(FLTaStringConstant, @"stringConstantValue");
+  XCTAssertEqual(FLTanIntConstant, 42);
+  XCTAssertEqual(FLTaDoubleConstant, 3.14);
+  XCTAssertEqual(FLTaBoolConstant, YES);
+}
+
 @end