+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 27.2.0
+
+* Adds support for sharing constants across platforms.
+
 ## 27.1.0
 
 * [swift] Adds `CaseIterable` conformance to generated enums.